add dtype in gym.spaces.Box

+++ environments/mypaint_env.py
@@ -42,8 +42,8 @@ class MyPaintEnv(gym.Env):
         # action space
         self.action_space = gym.spaces.Dict({
             'position': gym.spaces.Discrete(self.pos_resolution ** 2),
-            'pressure': gym.spaces.Box(low=0, high=1.0, shape=()),
-            'color': gym.spaces.Box(low=0, high=1.0, shape=(3,)),
+            'pressure': gym.spaces.Box(low=0, high=1.0, shape=(), dtype=float),
+            'color': gym.spaces.Box(low=0, high=1.0, shape=(3,), dtype=float),
             'prob': gym.spaces.Discrete(2)
         })
 
@@ -51,8 +51,8 @@ class MyPaintEnv(gym.Env):
         self.observation_space = gym.spaces.Dict({
             'image': gym.spaces.Box(low=0, high=255, shape=(self.imsize, self.imsize, 3), dtype=np.uint8),
             'position': gym.spaces.Discrete(self.pos_resolution ** 2),
-            'pressure': gym.spaces.Box(low=0, high=1.0, shape=()),
-            'color': gym.spaces.Box(low=0, high=1.0, shape=(3,)),
+            'pressure': gym.spaces.Box(low=0, high=1.0, shape=(), dtype=float),
+            'color': gym.spaces.Box(low=0, high=1.0, shape=(3,), dtype=float),
             'prob': gym.spaces.Discrete(1)
         })
